app/requirements.txt
uwsgi
Flask
mariadb
python-dotenv
bcrypt
uuid
PyCryptodome
flask-recaptcha